+++ app/src/components/home.jsx
@@ -5,10 +5,10 @@ import home from "../../styles/index.scss";
 import Header from "../containers/header";
 import Footer from "./footer";
 import FooterSecond from "./second_footer";
-import CaseStudySlider from "./case_study_slider";
-import CoverPage from "./cover_page";
-import InfoMap from "./info_map";
-import SuccessStoryGallery from "./success_story_gallery";
+import CaseStudySlider from "./home/case_study_slider";
+import CoverPage from "./home/cover_page";
+import InfoMap from "./home/info_map";
+import SuccessStoryGallery from "./home/success_story_gallery";
 
 class Home extends Component {
