+++ app/src/components/home/cover_page.jsx
@@ -1,10 +1,10 @@
 'use strict';
 import React, {Component} from "react";
 import $ from "jquery";
-import ButtonBoxSlider from "./button_box_slider";
-import cover_page from "../../styles/components/c_cover_page.scss";
-import boxtriangle from "../../assets/icons/box_triangle.svg";
-import box_triangle from "../../styles/components/c_box_triangle.scss";
+import ButtonBoxSlider from "../button_box_slider";
+import cover_page from "../../../styles/components/c_cover_page.scss";
+import boxtriangle from "../../../assets/icons/box_triangle.svg";
+import box_triangle from "../../../styles/components/c_box_triangle.scss";
 
 class CoverPage extends Component {
   gosection(){
